refactor: refactored file naming
All files were renamed, and also was fixed some code smells
