+++ .sonar-project.properties
@@ -8,7 +8,7 @@ sonar.sourceEncoding=UTF-8
 sonar.projectBaseDir=.
 
 sonar.sources="src"
-sonar.exclusions="**/services/*.spec.ts", "src/config"
+sonar.exclusions="**/*.spec.ts", "src/config", "src/shared/infra/typeorm"
 
 sonar.tests.inclusions="**/services/*.spec.ts"
 
@@ -17,5 +17,5 @@ sonar.host.url=http://localhost:9000
 sonar.login=9fb0ad5f757cc1b931c16516a5d5ebd69718806d
 
 sonar.javascript.coveragePlugin=lcov
-sonar.javascript.lcov.reportPaths=coverage/lcov.info
-
+sonar.javascript.lcov.reportPaths="coverage/lcov.info"
+sonar.testExecutionReportPaths="coverage/test-reporter.xml"
