+++ src/config/log.config.ts
@@ -1,5 +1,4 @@
 import axios from 'axios';
-import { get } from 'config';
 
 export default axios.create({
   baseURL: `http://${process.env.APM_SERVER_HOST}` || 'http://database:8080',
