+++ src/modules/tenants/repositories/ITenantsRepository.ts
@@ -1,5 +1,5 @@
-import CreateTenantDTO from '../dtos/create-tenant.dto';
-import Tenant from '../infra/typeorm/entities/tenant.entity';
+import CreateTenantDTO from '../dtos/CreateTenantDTO';
+import Tenant from '../infra/typeorm/entities/Tenant';
 
 export default interface ITenantsRepository {
   create(data: CreateTenantDTO): Promise<Tenant>;
