+++ src/modules/tenants/services/ListTenantService.spec.ts
@@ -1,9 +1,9 @@
 import { BadRequestException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
-import TenantsRepository from '../infra/typeorm/repositories/tenants.repository';
-import FakeTenantsRepository from '../repositories/fakes/fake-tenants.repository';
-import CreateTenantService from './create-tenant.service';
-import ListTenantService from './list-tenant.service';
+import TenantsRepository from '../infra/typeorm/repositories/TenantsRepository';
+import FakeTenantsRepository from '../repositories/fakes/FakeTenantsRepository';
+import CreateTenantService from './CreateTenantService';
+import ListTenantService from './ListTenantService';
 
 let listTenantService: ListTenantService;
 let createTenantService: CreateTenantService;
