+++ src/modules/tenants/services/ListTenantService.ts
@@ -1,7 +1,7 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { log_verbose } from '../../../shared/helper/app-logger';
-import Tenant from '../infra/typeorm/entities/tenant.entity';
-import TenantsRepository from '../infra/typeorm/repositories/tenants.repository';
+import Tenant from '../infra/typeorm/entities/Tenant';
+import TenantsRepository from '../infra/typeorm/repositories/TenantsRepository';
 
 @Injectable()
 export default class ListTenantService {
