+++ src/shared/exceptions/ErrorException.ts
